- Added bmp images and wavelet transform libraries - Added spreadsheet with lena_eye_16x16.bmp wavelet transform result - Added code for loading and wavelet transformation of input images

+++ trunk/wic/libwic/types.h
@@ -27,7 +27,7 @@ namespace wic {
 typedef unsigned char byte_t;
 
 //! \brief Тип действительных вейвлет коэффициентов
-typedef double w_t;
+typedef float w_t;
 
 //! \brief Тип целочисленных вейвлет коэффициентов
 typedef int wk_t;
